Copy the path string for parent directories creation
Previously, the display used the file path from the thread info structure,
which was actually mutated by the code which creates parent directories.  This
led the viewer tool to show multiple threades seemingly migrating the same
files or directories, which was erroneous.  Now, the string is first
duplicated, and the duplicate is manipulated for those operations.

Also, rename the error labels to be consistent throughout the file.

WIP for issue #24