Update thread data before the display can access it
The data was previously altered in-between the selection of the next file to
migrate and its migration, with a small unlocked period. This led the display
loop to potentially send information about the previous file while the thread
was on a new file to migrate. This change fixes this behavior.

WIP for issue #24