Fix an issue with the status storage
Previously, a status bucket name alone led to have the store path in the form
"name:/", which led before this fix to have an anonymous directory created on
an s3 backend (name://), which was treated differently from his parent. This
change fixes that by avoiding the use of double slashes if the status store is
a bucket and not a directory.