Fix the path used for the source deletions
Following the fixes over the filestate entries, and the way the paths are
managed, the source deletion code was not properly updated, leaving a wrong
path to remove, not including the bucket, which led to the DNS Lookup failing
(a DNS shall not start by a dot). The code is now using the proper src_path
field of the filestate entry, which includes the bucket name, and fixes the
issue observed.

Fixes issue #46